refactor: restructure backend to infrastructure/ + di/ layers, move frontend app/ to src/app/
Backend:
- Group connectors/, repository/, database.py under infrastructure/ parent layer
  following onion architecture (infrastructure is the outer layer boundary)
- Create di/ directory with domain-segmented DI files (bug, milestone,
  integration, workflow_column, user) instead of inline _get_usecase() functions
- Update all import paths across presentation/, usecase/, and main.py
- Presentation layer now imports from di/ — no direct infrastructure wiring

Frontend:
- Move app/ → src/app/ to consolidate all source under src/
  (Next.js supports src/app/ natively, no config change needed)
