+++ backend/app/di/user.py
@@ -1,0 +1,9 @@
+from fastapi import Depends
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from app.infrastructure.db.database import get_session
+from app.infrastructure.repository.user_repo import UserRepository
+
+
+def get_user_repo(session: AsyncSession = Depends(get_session)) -> UserRepository:
+    return UserRepository(session)
