+++ backend/app/infrastructure/connectors/base.py
