+++ backend/app/infrastructure/connectors/slack/verify.py
