+++ backend/app/presentation/integrations.py
@@ -1,9 +1,8 @@
 from typing import Any
 
 from fastapi import APIRouter, Depends, HTTPException, Query
-from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.database import get_session
+from app.di.integration import get_integration_usecase
 from app.domain.schemas import (
     IntegrationCreate,
     IntegrationEventResponse,
@@ -12,25 +11,20 @@ from app.domain.schemas import (
     SchemaFetchRequest,
     TestConnectionResponse,
 )
-from app.repository.integration_repo import IntegrationRepository
 from app.usecase.integration_usecase import IntegrationNotFoundError, IntegrationUsecase
 
 router = APIRouter(prefix="/api/integrations", tags=["integrations"])
 
 
-def _get_usecase(session: AsyncSession = Depends(get_session)) -> IntegrationUsecase:
-    return IntegrationUsecase(IntegrationRepository(session))
-
-
 @router.get("", response_model=list[IntegrationResponse])
-async def list_integrations(usecase: IntegrationUsecase = Depends(_get_usecase)) -> list[IntegrationResponse]:
+async def list_integrations(usecase: IntegrationUsecase = Depends(get_integration_usecase)) -> list[IntegrationResponse]:
     return await usecase.list_all()
 
 
 @router.post("", response_model=IntegrationResponse, status_code=201)
 async def create_integration(
     data: IntegrationCreate,
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> IntegrationResponse:
     return await usecase.create(data)
 
@@ -40,7 +34,7 @@ async def create_integration(
 async def get_source_schema(
     source_type: str,
     body: SchemaFetchRequest,
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> list[dict[str, Any]]:
     """Return available source fields for field mapping UI (uses provided credentials, not saved)."""
     return await usecase.fetch_schema(source_type, body.credentials)
@@ -49,7 +43,7 @@ async def get_source_schema(
 @router.get("/{integration_id}", response_model=IntegrationResponse)
 async def get_integration(
     integration_id: str,
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> IntegrationResponse:
     try:
         return await usecase.get(integration_id)
@@ -61,7 +55,7 @@ async def get_integration(
 async def update_integration(
     integration_id: str,
     data: IntegrationUpdate,
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> IntegrationResponse:
     try:
         return await usecase.update(integration_id, data)
@@ -72,7 +66,7 @@ async def update_integration(
 @router.delete("/{integration_id}", status_code=204)
 async def delete_integration(
     integration_id: str,
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> None:
     try:
         await usecase.delete(integration_id)
@@ -83,7 +77,7 @@ async def delete_integration(
 @router.post("/{integration_id}/test", response_model=TestConnectionResponse)
 async def test_integration(
     integration_id: str,
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> TestConnectionResponse:
     try:
         ok = await usecase.test_connection(integration_id)
@@ -96,7 +90,7 @@ async def test_integration(
 async def list_integration_events(
     integration_id: str,
     limit: int = Query(50, ge=1, le=200),
-    usecase: IntegrationUsecase = Depends(_get_usecase),
+    usecase: IntegrationUsecase = Depends(get_integration_usecase),
 ) -> list[IntegrationEventResponse]:
     try:
         return await usecase.list_events(integration_id, limit)
