+++ backend/app/presentation/milestones.py
@@ -1,32 +1,23 @@
 from fastapi import APIRouter, Depends, HTTPException
-from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.database import get_session
+from app.di.milestone import get_milestone_usecase
 from app.domain.schemas import MilestoneCreate, MilestoneResponse, MilestoneUpdate
-from app.repository.milestone_repo import MilestoneRepository
 from app.usecase.milestone_usecase import MilestoneNotFoundError, MilestoneUsecase
 
 router = APIRouter(prefix="/api/milestones", tags=["milestones"])
 
 
-def _get_usecase(
-    session: AsyncSession = Depends(get_session),
-) -> MilestoneUsecase:
-    repo = MilestoneRepository(session)
-    return MilestoneUsecase(repo)
-
-
 @router.post("", response_model=MilestoneResponse, status_code=201)
 async def create_milestone(
     data: MilestoneCreate,
-    usecase: MilestoneUsecase = Depends(_get_usecase),
+    usecase: MilestoneUsecase = Depends(get_milestone_usecase),
 ) -> MilestoneResponse:
     return await usecase.create(data)
 
 
 @router.get("", response_model=list[MilestoneResponse])
 async def list_milestones(
-    usecase: MilestoneUsecase = Depends(_get_usecase),
+    usecase: MilestoneUsecase = Depends(get_milestone_usecase),
 ) -> list[MilestoneResponse]:
     return await usecase.list_all()
 
@@ -34,7 +25,7 @@ async def list_milestones(
 @router.get("/{milestone_id}", response_model=MilestoneResponse)
 async def get_milestone(
     milestone_id: str,
-    usecase: MilestoneUsecase = Depends(_get_usecase),
+    usecase: MilestoneUsecase = Depends(get_milestone_usecase),
 ) -> MilestoneResponse:
     try:
         return await usecase.get(milestone_id)
@@ -46,7 +37,7 @@ async def get_milestone(
 async def update_milestone(
     milestone_id: str,
     data: MilestoneUpdate,
-    usecase: MilestoneUsecase = Depends(_get_usecase),
+    usecase: MilestoneUsecase = Depends(get_milestone_usecase),
 ) -> MilestoneResponse:
     try:
         return await usecase.update(milestone_id, data)
@@ -57,7 +48,7 @@ async def update_milestone(
 @router.delete("/{milestone_id}", status_code=204)
 async def delete_milestone(
     milestone_id: str,
-    usecase: MilestoneUsecase = Depends(_get_usecase),
+    usecase: MilestoneUsecase = Depends(get_milestone_usecase),
 ) -> None:
     try:
         await usecase.delete(milestone_id)
