+++ frontend/src/app/analytics/page.tsx
