+++ frontend/src/app/board/page.tsx
