+++ frontend/src/app/bugs/[id]/page.tsx
